requirements.txt
PyMuPDF
google-generativeai
pydantic
